lib: Rework select
Support select FDs mixed with user FDs and system FDs.
If user FD is not connected(ingress flow not applied), we
still consider it as system FD because the traffic of the FD
is still processed in kernel.
However, select mixed FDs impacts performance, user should
avoid this in datapath processing.

Signed-off-by: Jun Yang <jun.yang@nxp.com>